Remove specific classes. Update TextBox element with filter option.
