--- ponysdk/src-core/main/java/com/ponysdk/ui/server/basic/PFilterTextBox.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (c) 2011 PonySDK
- *  Owners:
- *  Luciano Broussal  <luciano.broussal AT gmail.com>
- *  Mathieu Barbier   <mathieu.barbier AT gmail.com>
- *  Nicolas Ciaravola <nicolas.ciaravola.pro AT gmail.com>
- *  
- *  WebSite:
- *  http://code.google.com/p/pony-sdk/
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package com.ponysdk.ui.server.basic;
-
-import com.ponysdk.core.instruction.Update;
-import com.ponysdk.core.stm.Txn;
-import com.ponysdk.ui.terminal.Dictionnary.PROPERTY;
-import com.ponysdk.ui.terminal.WidgetType;
-
-public class PFilterTextBox extends PTextBox {
-
-    /**
-     * Apply a filter to the textbox. Value set to the textbox will be filtered. <br>
-     * Example: ([a-zA-Z0-9])
-     * 
-     * @param regExp
-     *            the regular expression to use as filter.
-     */
-    public void setFilter(final String regExp) {
-        final Update update = new Update(getID());
-        update.put(PROPERTY.FILTER, regExp);
-        Txn.get().getTxnContext().save(update);
-    }
-
-    @Override
-    protected WidgetType getWidgetType() {
-        return WidgetType.FILTER_TEXTBOX;
-    }
-
-}
