+++ ponysdk/src-core/main/java/com/ponysdk/ui/server/basic/PTextBox.java
@@ -115,4 +115,17 @@ public class PTextBox extends PTextBoxBase {
         Txn.get().getTxnContext().save(update);
     }
 
+    /**
+     * Sets a filter to the textbox. Value sets to the textbox will be filtered. <br>
+     * Example: ([a-zA-Z0-9])
+     * 
+     * @param regExp
+     *            the regular expression to use as filter.
+     */
+    public void setFilter(final String regExp) {
+        final Update update = new Update(getID());
+        update.put(PROPERTY.FILTER, regExp);
+        Txn.get().getTxnContext().save(update);
+    }
+
 }
