+++ ponysdk/src-core/main/java/com/ponysdk/ui/terminal/WidgetType.java
@@ -47,7 +47,6 @@ public enum WidgetType {
     NUMBERBOX,
     TEXTBOX,
     PASSWORD_TEXTBOX,
-    FILTER_TEXTBOX,
     LISTBOX,
     IMAGE,
     ANCHOR,
