+++ ponysdk/src-core/main/resources/spec/propertiesDictionnary.json
@@ -262,6 +262,7 @@
 		"maxLength": [],
 		"year": [],
 		"month": [],
-		"day": []
+		"day": [],
+		"filter": []
 	}
 }
